requirements.txt
zhipuai
transformers
torchvision
tqdm
Pillow
numpy
google-genai
pydub
scipy
httpx
openai